04-AI-Resume-Ranker/requirements.txt
spacy
scikit-learn
pandas
pdfminer.six
flask